research/storage_research/src/requirements.txt
pandas
pydantic
python-dotenv
psycopg2-binary
Faker
pymongo
matplotlib